feat: core entities account refactor
