+++ packages/core-entities/src/client/account-registry-client.ts
@@ -1,47 +1,32 @@
 import {
-  ERC20ABI,
-  NotionalV3,
-  NotionalV3ABI,
-} from '@notional-finance/contracts';
-import {
-  AggregateCall,
-  NO_OP,
-  getMulticall,
-} from '@notional-finance/multicall';
-import {
-  encodefCashId,
   filterEmpty,
   getProviderFromNetwork,
-  INTERNAL_TOKEN_PRECISION,
-  MAX_APPROVAL,
   Network,
-  NotionalAddress,
-  PRIME_CASH_VAULT_MATURITY,
-  RATE_PRECISION,
-  SCALAR_PRECISION,
-  ZERO_ADDRESS,
 } from '@notional-finance/util';
-import { BigNumber, Contract } from 'ethers';
-import { BehaviorSubject, filter, lastValueFrom, take, takeUntil } from 'rxjs';
+import { providers } from 'ethers';
+import { BehaviorSubject, from, of, switchMap } from 'rxjs';
 import {
   AccountDefinition,
   BalanceStatement,
   CacheSchema,
-  Registry,
   TokenBalance,
   AccountHistory,
-  AccountIncentiveDebt,
 } from '..';
 import { Routes } from '../server';
 import {
   fetchGraph,
   fetchUsingGraph,
-  fetchUsingMulticall,
   loadGraphClientDeferred,
 } from '../server/server-registry';
 import { ClientRegistry } from './client-registry';
 // eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
 import { BalanceSnapshot, ProfitLossLineItem, Token } from '../.graphclient';
+import {
+  parseBalanceStatement,
+  parseCurrentBalanceStatement,
+} from './accounts/balance-statement';
+import { parseTransactionHistory } from './accounts/transaction-history';
+import { fetchCurrentAccount } from './accounts/current-account';
 
 export enum AccountFetchMode {
   // Used for the frontend UI, will fetch data for a single account direct from
@@ -55,8 +40,14 @@ export enum AccountFetchMode {
 export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
   protected _activeAccount = new BehaviorSubject<string | null>(null);
 
+  protected _walletProvider?: providers.Provider;
+
   constructor(cacheHostname: string, public fetchMode: AccountFetchMode) {
     super(cacheHostname);
+  }
+
+  set walletProvider(p: providers.Provider) {
+    this._walletProvider = p;
   }
 
   protected cachePath() {
@@ -89,46 +80,51 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
     return this.getLatestFromSubject(network, account);
   }
 
-  public subscribeAccount(network: Network, account: string) {
-    const sub = this.subscribeSubject(network, account);
-    if (!sub) throw Error(`Account ${account} on ${network} not found`);
+  public subscribeActiveAccount(network: Network) {
+    return this.activeAccount$.pipe(
+      switchMap((a) =>
+        a
+          ? from(
+              new Promise<ReturnType<typeof this.subscribeSubject>>((resolve) =>
+                this.onSubjectKeyRegistered(network, a, () =>
+                  resolve(this.subscribeSubject(network, a))
+                )
+              )
+            )
+          : of(of(null))
+      ),
+      filterEmpty(),
+      switchMap((o) => o)
+    );
+  }
 
-    if (this.fetchMode === AccountFetchMode.SINGLE_ACCOUNT_DIRECT) {
-      if (account !== this.activeAccount)
-        throw Error('Can only fetch active account');
+  public async refreshBalanceHistory(network: Network) {
+    if (this.activeAccount === null) return undefined;
+    const subject = this._getNetworkSubjects(network).get(this.activeAccount);
+    const currentAccount = subject?.value;
 
-      // Take the values from the subscription until the active account changes
-      return sub.pipe(
-        takeUntil(this.activeAccount$.pipe(filter((a) => a !== account))),
-        filterEmpty()
+    if (currentAccount) {
+      const newAccount = await this._refreshBalanceHistory(
+        network,
+        this.activeAccount,
+        Object.assign(currentAccount)
       );
+      subject.next(newAccount);
+
+      return newAccount;
     }
 
-    return sub;
+    return undefined;
   }
 
   /** Triggers a manual refresh of the active account and provides an optional callback on refresh complete */
   public refreshActiveAccount(network: Network) {
     if (this.activeAccount === null) throw Error('No active account');
-
-    // Returns a promise for the  value from the manual refresh
-    const p = lastValueFrom(
-      this.subscribeAccount(network, this.activeAccount).pipe(take(1))
-    );
-
     this.triggerRefresh(network);
-
-    return p;
   }
 
-  /** Switches the actively listened account to the newly registered one on the specified network*/
-  public onAccountReady(
-    network: Network,
-    account: string,
-    fn: (a: AccountDefinition) => void
-  ) {
-    this.onSubjectKeyReady(network, account, fn);
-
+  /** Switches the actively listened account to the newly registered one */
+  public setAccount(network: Network, account: string) {
     if (
       this.fetchMode === AccountFetchMode.SINGLE_ACCOUNT_DIRECT &&
       this.activeAccount !== account
@@ -142,7 +138,9 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
 
       // Kick off a refresh of the accounts if in single account mode and we are
       // changing the account
-      this.triggerRefresh(network);
+      return this.triggerRefreshPromise(network);
+    } else {
+      return;
     }
   }
 
@@ -150,7 +148,10 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
     network: Network
   ): Promise<CacheSchema<AccountDefinition>> {
     if (this.fetchMode === AccountFetchMode.SINGLE_ACCOUNT_DIRECT) {
-      return this._fetchSingleAccount(network);
+      return this._fetchSingleAccount(
+        network,
+        this._walletProvider || getProviderFromNetwork(network)
+      );
     } else if (this.fetchMode === AccountFetchMode.BATCH_ACCOUNT_VIA_SERVER) {
       return this._fetchBatchAccounts(network);
     }
@@ -158,14 +159,27 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
     throw Error('Unknown fetch mode');
   }
 
-  private async _fetchSingleAccount(network: Network) {
-    const provider = getProviderFromNetwork(network);
-    const notional = new Contract(
-      NotionalAddress[network],
-      NotionalV3ABI,
-      provider
-    ) as NotionalV3;
+  private async _refreshBalanceHistory(
+    network: Network,
+    address: string,
+    account: AccountDefinition
+  ) {
+    const [txnHistory, balanceStatements] = await Promise.all([
+      this.fetchTransactionHistory(network, address),
+      this.fetchBalanceStatements(network, address),
+    ]);
 
+    // Set the balance statement and txn history
+    account.balanceStatement = balanceStatements.finalResults[address];
+    account.accountHistory = txnHistory.finalResults[address];
+
+    return account;
+  }
+
+  private async _fetchSingleAccount(
+    network: Network,
+    provider: providers.Web3Provider
+  ) {
     if (this.activeAccount === null) {
       return {
         values: [],
@@ -174,353 +188,19 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
         lastUpdateBlock: this.getLastUpdateBlock(network),
       };
     }
-    const activeAccount = this.activeAccount;
-    const tokens = Registry.getTokenRegistry();
-    const config = Registry.getConfigurationRegistry();
-    const walletTokensToTrack = tokens
-      .getAllTokens(network)
-      .filter((t) => t.currencyId !== undefined && t.tokenType === 'Underlying')
-      .map((t) => t.address);
-
-    const walletCalls = walletTokensToTrack.flatMap<AggregateCall>(
-      (address) => {
-        const def = tokens.getTokenByAddress(network, address);
-        if (def.address === ZERO_ADDRESS) {
-          return [
-            {
-              stage: 0,
-              target: getMulticall(provider),
-              method: 'getEthBalance',
-              args: [activeAccount],
-              key: `${address}.balance`,
-              transform: (b: BigNumber) => {
-                return { balances: [TokenBalance.from(b, def)] };
-              },
-            },
-            {
-              stage: 0,
-              target: NO_OP,
-              method: NO_OP,
-              key: `${address}.allowance`,
-              transform: () => {
-                return TokenBalance.from(MAX_APPROVAL, def);
-              },
-            },
-          ];
-        } else {
-          return [
-            {
-              stage: 0,
-              target: new Contract(address, ERC20ABI, provider),
-              method: 'balanceOf',
-              args: [activeAccount],
-              key: `${address}.balance`,
-              transform: (b: BigNumber) => {
-                return { balances: [TokenBalance.from(b, def)] };
-              },
-            },
-            {
-              stage: 0,
-              target: new Contract(address, ERC20ABI, provider),
-              method: 'allowance',
-              args: [activeAccount, notional.address],
-              key: `${address}.allowance`,
-              transform: (b: BigNumber) => {
-                return TokenBalance.from(b, def);
-              },
-            },
-          ];
-        }
-      }
-    );
-
-    const secondaryIncentiveCalls = tokens
-      .getAllTokens(network)
-      .filter(
-        (t) =>
-          t.currencyId !== undefined &&
-          t.tokenType === 'nToken' &&
-          config.getSecondaryRewarder(t)
-      )
-      .flatMap((t) => {
-        const rewarder = config.getSecondaryRewarder(t);
-        const secondary = config.getAnnualizedSecondaryIncentives(t);
-        if (!rewarder || !secondary) return [];
-        const { rewardToken } = secondary;
-        const rewardPrecision = BigNumber.from(10).pow(rewardToken.decimals);
-        return [
-          {
-            stage: 0,
-            target: rewarder,
-            method: 'rewardDebtPerAccount',
-            args: [this.activeAccount],
-            key: `${t.currencyId}.secondaryIncentiveDebt`,
-            transform: (r: BigNumber) => ({
-              // Secondary rewarder always returns this in 18 decimals
-              value: TokenBalance.from(r, rewardToken).scale(
-                rewardPrecision,
-                SCALAR_PRECISION
-              ),
-              currencyId: t.currencyId,
-            }),
-          },
-        ];
-      });
-
-    const vaultCalls =
-      config.getAllListedVaults(network)?.flatMap<AggregateCall>((v) => {
-        return [
-          {
-            stage: 0,
-            target: notional,
-            method: 'getVaultAccount',
-            args: [this.activeAccount, v.vaultAddress],
-            key: `${v.vaultAddress}.balance`,
-            transform: (
-              vaultAccount: Awaited<ReturnType<NotionalV3['getVaultAccount']>>
-            ) => {
-              const maturity = vaultAccount.maturity.toNumber();
-              if (maturity === 0) return { balances: [] };
-              const {
-                vaultShareID,
-                primaryDebtID,
-                primaryCashID,
-                primaryTokenId,
-              } = config.getVaultIDs(network, v.vaultAddress, maturity);
-              const balances = [
-                TokenBalance.fromID(
-                  vaultAccount.vaultShares,
-                  vaultShareID,
-                  network
-                ),
-                this._parseVaultDebtBalance(
-                  primaryDebtID,
-                  primaryTokenId,
-                  vaultAccount.accountDebtUnderlying,
-                  maturity,
-                  network
-                ),
-              ];
-
-              if (!vaultAccount.tempCashBalance.isZero()) {
-                balances.push(
-                  TokenBalance.fromID(
-                    vaultAccount.tempCashBalance,
-                    primaryCashID,
-                    network
-                  )
-                );
-              }
-
-              return {
-                balances,
-                vaultLastUpdateTime: {
-                  [v.vaultAddress]: vaultAccount.lastUpdateBlockTime.toNumber(),
-                },
-              };
-            },
-          },
-          {
-            stage: 0,
-            target: notional,
-            method: 'getVaultAccountSecondaryDebt',
-            args: [this.activeAccount, v.vaultAddress],
-            key: `${v.vaultAddress}.balance2`,
-            transform: (
-              r: Awaited<ReturnType<NotionalV3['getVaultAccountSecondaryDebt']>>
-            ) => {
-              const maturity = r.maturity.toNumber();
-              if (maturity === 0) return { balances: [] };
-              const {
-                secondaryOneCashID,
-                secondaryOneDebtID,
-                secondaryOneTokenId,
-                secondaryTwoCashID,
-                secondaryTwoDebtID,
-                secondaryTwoTokenId,
-              } = config.getVaultIDs(network, v.vaultAddress, maturity);
-
-              const secondaries: TokenBalance[] = [];
-
-              if (
-                secondaryOneDebtID &&
-                secondaryOneTokenId &&
-                !r.accountSecondaryDebt[0].isZero()
-              ) {
-                secondaries.push(
-                  this._parseVaultDebtBalance(
-                    secondaryOneDebtID,
-                    secondaryOneTokenId,
-                    r.accountSecondaryDebt[0],
-                    maturity,
-                    network
-                  )
-                );
-              }
-
-              if (
-                secondaryTwoDebtID &&
-                secondaryTwoTokenId &&
-                !r.accountSecondaryDebt[1].isZero()
-              ) {
-                secondaries.push(
-                  this._parseVaultDebtBalance(
-                    secondaryTwoDebtID,
-                    secondaryTwoTokenId,
-                    r.accountSecondaryDebt[1],
-                    maturity,
-                    network
-                  )
-                );
-              }
-
-              if (
-                secondaryOneCashID &&
-                !r.accountSecondaryCashHeld[0].isZero()
-              ) {
-                secondaries.push(
-                  TokenBalance.fromID(
-                    r.accountSecondaryCashHeld[0],
-                    secondaryOneCashID,
-                    network
-                  )
-                );
-              }
-
-              if (
-                secondaryTwoCashID &&
-                !r.accountSecondaryCashHeld[1].isZero()
-              ) {
-                secondaries.push(
-                  TokenBalance.fromID(
-                    r.accountSecondaryCashHeld[1],
-                    secondaryTwoCashID,
-                    network
-                  )
-                );
-              }
-
-              return { balances: secondaries };
-            },
-          },
-        ];
-      }) || [];
-
-    const NOTE = Registry.getTokenRegistry().getTokenBySymbol(network, 'NOTE');
-    const calls = [
-      ...walletCalls,
-      ...vaultCalls,
-      ...secondaryIncentiveCalls,
-      {
-        target: notional,
-        method: 'getAccount',
-        args: [this.activeAccount],
-        key: `${notional.address}.account`,
-        transform: (r: Awaited<ReturnType<NotionalV3['getAccount']>>) => {
-          const accountIncentiveDebt: AccountIncentiveDebt[] = [];
-
-          const accountBalances = r.accountBalances.flatMap((b) => {
-            const balances: TokenBalance[] = [];
-
-            if (b.cashBalance.gt(0)) {
-              const pCash = tokens.getPrimeCash(network, b.currencyId);
-              balances.push(TokenBalance.from(b.cashBalance, pCash));
-            } else if (b.cashBalance.lt(0)) {
-              const pCash = tokens.getPrimeCash(network, b.currencyId);
-              const pDebt = tokens.getPrimeDebt(network, b.currencyId);
-              balances.push(
-                TokenBalance.from(b.cashBalance, pCash).toToken(pDebt)
-              );
-            }
-
-            if (b.nTokenBalance.gt(0)) {
-              const nToken = tokens.getNToken(network, b.currencyId);
-              balances.push(TokenBalance.from(b.nTokenBalance, nToken));
-            }
-
-            accountIncentiveDebt.push({
-              value: TokenBalance.from(b.accountIncentiveDebt, NOTE),
-              currencyId: b.currencyId,
-            });
-
-            return balances;
-          });
-
-          const portfolioBalances = r.portfolio.map((a) => {
-            const fCashId = encodefCashId(a.currencyId, a.maturity.toNumber());
-            return TokenBalance.fromID(a.notional, fCashId, network);
-          });
-
-          return {
-            balances: accountBalances.concat(portfolioBalances),
-            accountIncentiveDebt,
-            allowPrimeBorrow: r.accountContext.allowPrimeBorrow,
-          };
-        },
-      },
-    ];
-
-    const { finalResults: balanceStatement } =
-      await this.fetchBalanceStatements(network, activeAccount);
-    const { finalResults: accountHistory } = await this.fetchTransactionHistory(
+    const account = await fetchCurrentAccount(
       network,
-      activeAccount
+      this.activeAccount,
+      provider
     );
-    return fetchUsingMulticall<AccountDefinition>(network, calls, [
-      (results: Record<string, unknown>) => {
-        return {
-          [activeAccount]: {
-            address: activeAccount,
-            network,
-            balanceStatement: balanceStatement[activeAccount],
-            accountHistory: accountHistory[activeAccount],
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            allowPrimeBorrow: (results[`${notional.address}.account`] as any)[
-              'allowPrimeBorrow'
-            ],
-            balances: Object.keys(results).flatMap((k) =>
-              k.includes('balance') || k.includes('account')
-                ? // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                  ((results[k] as any)['balances'] as TokenBalance[])
-                : []
-            ),
-            accountIncentiveDebt: Object.keys(results).flatMap(
-              (k) =>
-                (k.includes('account')
-                  ? // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    (results[k] as any)['accountIncentiveDebt']
-                  : []) as AccountIncentiveDebt[]
-            ),
-            secondaryIncentiveDebt: Object.keys(results).flatMap(
-              (k) =>
-                (k.includes('secondaryIncentiveDebt')
-                  ? results[k]
-                  : []) as AccountIncentiveDebt[]
-            ),
-            vaultLastUpdateTime: Object.keys(results).reduce(
-              (agg, k) =>
-                Object.assign(
-                  agg,
-                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                  ((results[k] as any)['vaultLastUpdateTime'] as Record<
-                    string,
-                    number
-                  >) || {}
-                ),
-              {} as Record<string, number>
-            ),
-            allowances: Object.keys(results)
-              .filter((k) => k.includes('.allowance'))
-              .map((k) => {
-                return {
-                  spender: notional.address,
-                  amount: results[k] as TokenBalance,
-                };
-              }),
-          },
-        };
-      },
-    ]);
+
+    account.values[0][1] = await this._refreshBalanceHistory(
+      network,
+      this.activeAccount,
+      account.values[0][1] as AccountDefinition
+    );
+
+    return account;
   }
 
   public async fetchTransactionHistory(network: Network, account: string) {
@@ -534,7 +214,7 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
           [account]:
             r.account?.profitLossLineItems
               ?.map((p) => {
-                return this._parseTransactionHistory(
+                return parseTransactionHistory(
                   p as ProfitLossLineItem,
                   network
                 );
@@ -561,14 +241,15 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
               .map(({ current, snapshots, token }) => {
                 if (!token.underlying) throw Error('Unknown underlying');
                 return {
-                  ...this._parseCurrentBalanceStatement(
+                  ...parseCurrentBalanceStatement(
                     current as BalanceSnapshot,
                     token as Token,
                     network
                   ),
+                  // NOTE: this is used to populate the account history chart in the portfolio page
                   historicalSnapshots:
                     snapshots?.map((s) =>
-                      this._parseBalanceStatement(
+                      parseBalanceStatement(
                         token.id,
                         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
                         token.underlying!.id,
@@ -584,26 +265,6 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
         accountId: account.toLowerCase(),
       }
     );
-  }
-
-  private _parseVaultDebtBalance(
-    debtID: string,
-    underlyingID: string,
-    balance: BigNumber,
-    maturity: number,
-    network: Network
-  ) {
-    if (maturity === PRIME_CASH_VAULT_MATURITY) {
-      // In in the prime vault maturity, convert from underlying back to prime debt denomination
-      const tokens = Registry.getTokenRegistry();
-      const vaultDebtToken = tokens.getTokenByID(network, debtID);
-      const pDebt = TokenBalance.fromID(balance, underlyingID, network)
-        .scaleFromInternal()
-        .toToken(tokens.unwrapVaultToken(vaultDebtToken));
-      return TokenBalance.fromID(pDebt.n, debtID, network);
-    }
-
-    return TokenBalance.fromID(balance, debtID, network);
   }
 
   private async _fetchBatchAccounts(network: Network) {
@@ -628,7 +289,7 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
             balanceStatement: a.balances
               ?.filter((b) => !!b.token.underlying)
               .map((b) =>
-                this._parseCurrentBalanceStatement(
+                parseCurrentBalanceStatement(
                   b.current as BalanceSnapshot,
                   b.token as Token,
                   network
@@ -636,7 +297,7 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
               ),
             accountHistory:
               a.profitLossLineItems?.map((p) =>
-                this._parseTransactionHistory(p as ProfitLossLineItem, network)
+                parseTransactionHistory(p as ProfitLossLineItem, network)
               ) || [],
           } as AccountDefinition;
 
@@ -647,205 +308,6 @@ export class AccountRegistryClient extends ClientRegistry<AccountDefinition> {
       'accounts'
     );
 
-    // NOTE claims are turned off while not running the contest
-    // const provider = getProviderFromNetwork(network, !USE_CROSS_FETCH);
-    // const nowSeconds = getNowSeconds();
-    // const notional = new Contract(
-    //   NotionalAddress[network],
-    //   NotionalV3ABI,
-    //   provider
-    // ) as NotionalV3;
-
-    // const { results: noteClaims } = await aggregate<TokenBalance>(
-    //   accounts.values.map(([a]) => {
-    //     return {
-    //       target: notional,
-    //       method: 'nTokenGetClaimableIncentives',
-    //       args: [a, nowSeconds + 100],
-    //       key: a,
-    //       transform: (
-    //         r: Awaited<ReturnType<NotionalV3['nTokenGetClaimableIncentives']>>
-    //       ) => {
-    //         return TokenBalance.fromID(r, 'NOTE', Network.All);
-    //       },
-    //     };
-    //   }),
-    //   provider
-    // );
-
-    // Object.entries(noteClaims).forEach(([acct, currentNOTE]) => {
-    //   const definition = accounts.values.find(([a]) => a === acct);
-    //   if (definition && definition[1]) {
-    //     definition[1].noteClaim = {
-    //       currentNOTE,
-    //       noteAccruedPerSecond: currentNOTE.copy(0),
-    //     };
-    //   }
-    // });
-
     return accounts;
-  }
-
-  private _parseCurrentBalanceStatement(
-    current: BalanceSnapshot,
-    token: Token,
-    network: Network
-  ) {
-    const tokens = Registry.getTokenRegistry();
-    if (!token.underlying) throw Error('Unknown underlying');
-    const tokenId = token.id;
-    const underlying = token.underlying;
-    const currentStatement = this._parseBalanceStatement(
-      tokenId,
-      underlying.id,
-      current as BalanceSnapshot,
-      network
-    );
-
-    const currentProfitAndLoss = currentStatement.balance
-      .toUnderlying()
-      .sub(
-        currentStatement.adjustedCostBasis.scale(
-          currentStatement.balance.n,
-          INTERNAL_TOKEN_PRECISION
-        )
-      );
-    const totalInterestAccrual = currentProfitAndLoss.sub(
-      currentStatement.totalILAndFees
-    );
-    const incentives =
-      current.incentives?.map((i) => ({
-        adjustedClaimed: TokenBalance.fromSymbol(
-          i.adjustedClaimed,
-          i.rewardToken.symbol,
-          network
-        ),
-        totalClaimed: TokenBalance.fromSymbol(
-          i.totalClaimed,
-          i.rewardToken.symbol,
-          network
-        ),
-      })) || [];
-
-    return {
-      token: tokens.getTokenByID(network, tokenId),
-      blockNumber: current.blockNumber,
-      underlying: tokens.getTokenByID(network, underlying.id),
-      currentBalance: currentStatement.balance,
-      adjustedCostBasis: currentStatement.adjustedCostBasis,
-      totalILAndFees: currentStatement.totalILAndFees,
-      impliedFixedRate: currentStatement.impliedFixedRate,
-      totalProfitAndLoss: currentProfitAndLoss,
-      totalInterestAccrual,
-      accumulatedCostRealized: currentStatement.accumulatedCostRealized,
-      incentives,
-    };
-  }
-
-  private _parseBalanceStatement(
-    tokenId: string,
-    underlyingId: string,
-    snapshot: BalanceSnapshot,
-    network: Network
-  ) {
-    const balance = TokenBalance.fromID(
-      snapshot.currentBalance,
-      tokenId,
-      network
-    );
-
-    const adjustedCostBasis = TokenBalance.fromID(
-      snapshot.adjustedCostBasis,
-      underlyingId,
-      network
-    );
-    const accumulatedCostRealized = adjustedCostBasis.scale(
-      balance,
-      balance.precision
-    );
-    return {
-      balance,
-      adjustedCostBasis,
-      timestamp: snapshot.timestamp,
-      accumulatedCostRealized,
-      totalILAndFees: TokenBalance.fromID(
-        snapshot.totalILAndFeesAtSnapshot,
-        underlyingId,
-        network
-      ),
-      totalProfitAndLoss: TokenBalance.fromID(
-        snapshot.totalProfitAndLossAtSnapshot,
-        underlyingId,
-        network
-      ),
-      totalInterestAccrual: TokenBalance.fromID(
-        snapshot.totalInterestAccrualAtSnapshot,
-        underlyingId,
-        network
-      ),
-      impliedFixedRate: snapshot.impliedFixedRate
-        ? (snapshot.impliedFixedRate * 100) / RATE_PRECISION
-        : undefined,
-    };
-  }
-
-  private _parseTransactionHistory(p: ProfitLossLineItem, network: Network) {
-    const tokenId = p.token.id;
-    const underlyingId = p.underlyingToken.id;
-    const token = Registry.getTokenRegistry().getTokenByID(network, tokenId);
-    const vaultName =
-      !!token.vaultAddress && token.vaultAddress !== ZERO_ADDRESS
-        ? Registry.getConfigurationRegistry().getVaultName(
-            token.network,
-            token.vaultAddress
-          )
-        : undefined;
-
-    let tokenAmount = TokenBalance.fromID(p.tokenAmount, tokenId, network);
-    let underlyingAmountRealized = TokenBalance.fromID(
-      p.underlyingAmountRealized,
-      underlyingId,
-      network
-    );
-    let underlyingAmountSpot = TokenBalance.fromID(
-      p.underlyingAmountSpot,
-      underlyingId,
-      network
-    );
-
-    if (tokenAmount.tokenType === 'PrimeDebt') {
-      tokenAmount = tokenAmount.neg();
-      underlyingAmountRealized = underlyingAmountRealized.neg();
-      underlyingAmountSpot = underlyingAmountSpot.neg();
-    } else if (tokenAmount.tokenType === 'fCash' && token.isFCashDebt) {
-      underlyingAmountRealized = underlyingAmountRealized.neg();
-      underlyingAmountSpot = underlyingAmountSpot.neg();
-    }
-
-    return {
-      timestamp: p.timestamp,
-      blockNumber: p.blockNumber,
-      token,
-      vaultName,
-      underlying: Registry.getTokenRegistry().getTokenByID(
-        network,
-        underlyingId
-      ),
-      bundleName: p.bundle?.bundleName,
-      transactionHash: p.transactionHash.id,
-      tokenAmount,
-      underlyingAmountRealized,
-      underlyingAmountSpot,
-      realizedPrice: TokenBalance.fromID(
-        p.realizedPrice,
-        underlyingId,
-        network
-      ),
-      spotPrice: TokenBalance.fromID(p.spotPrice, underlyingId, network),
-      impliedFixedRate: p.impliedFixedRate
-        ? (p.impliedFixedRate * 100) / RATE_PRECISION
-        : undefined,
-      isTransientLineItem: p.isTransientLineItem,
-    };
   }
 }
